LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/geolocation-API/enabled-by-permissions-policy.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Permission policy 'Geolocation' check failed for document with origin 'https://127.0.0.1:9443'.


PASS Permissions-Policy header geolocation=* allows the top-level document.
PASS Permissions-Policy header geolocation=* allows same-origin iframes.
PASS Permissions-Policy header geolocation=* allows cross-origin iframes.
